add enable_qlora in train.py Summary:
Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
